Convereted other models (fibrosis and invasion) to use default cell function. Updated fibrosis to use unified cell-ECM interaction functions. Altered base ECM density remodeling rate as well.
