+++ custom_modules/AMIGOS-invasion_uncoupled.cpp
@@ -85,7 +85,7 @@ void create_cell_types( void )
 	   This is a good place to set default functions. 
 	*/ 
 	
-	create_default_ECM_compatible_agent(); // in cell_ECM_interactions.cpp - sets default cell type to have ECM interactions
+	create_default_ECM_compatible_agent(); // in cell_ECM_interactions.cpp. Sets custom velocity function (cell-ECM motility interaction) and custom cell rule (ECM remodeling).
 	cell_defaults.phenotype.motility.migration_speed = parameters.doubles("default_cell_speed");  //rwh
 	
 	/*
@@ -1382,13 +1382,6 @@ void chemotaxis_oxygen( Cell* pCell , Phenotype& phenotype , double dt )
 	return; 
 }
 
-void tumor_cell_phenotype_with_oncoprotein( Cell* pCell , Phenotype& phenotype , double dt ) 
-{	
-	// o2-based birth and death, nothing more 
-	update_cell_and_death_parameters_O2_based(pCell,phenotype,dt);
-	return; 
-}
-
 std::vector<std::string> AMIGOS_invasion_coloring_function( Cell* pCell )
 {
 	// leaders are blue 
@@ -1639,332 +1632,5 @@ void ecm_update_from_cell_velocity_vector(Cell* pCell , Phenotype& phenotype , d
     ecm.ecm_voxels[nearest_ecm_voxel_index].anisotropy = anisotropy + r_anisotropy * dt  * (1- anisotropy);
     
     // END Cell-ECM Anisotropy Modification
-    return;
-}
-
-void SVG_plot_custom( std::string filename , Microenvironment& M, double z_slice , double time, std::vector<std::string> (*cell_coloring_function)(Cell*), std::string line_pattern )
-{
-	double X_lower = M.mesh.bounding_box[0];
-	double X_upper = M.mesh.bounding_box[3];
- 
-	double Y_lower = M.mesh.bounding_box[1]; 
-	double Y_upper = M.mesh.bounding_box[4]; 
-
-	double plot_width = X_upper - X_lower; 
-	double plot_height = Y_upper - Y_lower; 
-
-	double font_size = 0.025 * plot_height; // PhysiCell_SVG_options.font_size; 
-	double top_margin = font_size*(.2+1+.2+.9+.5 ); 
-
-	// open the file, write a basic "header"
-	std::ofstream os( filename , std::ios::out );
-	if( os.fail() )
-	{ 
-		std::cout << std::endl << "Error: Failed to open " << filename << " for SVG writing." << std::endl << std::endl; 
-
-		std::cout << std::endl << "Error: We're not writing data like we expect. " << std::endl
-		<< "Check to make sure your save directory exists. " << std::endl << std::endl
-		<< "I'm going to exit with a crash code of -1 now until " << std::endl 
-		<< "you fix your directory. Sorry!" << std::endl << std::endl; 
-		exit(-1); 
-	} 
-	
-	Write_SVG_start( os, plot_width , plot_height + top_margin );
-
-	// draw the background 
-	Write_SVG_rect( os , 0 , 0 , plot_width, plot_height + top_margin , 0.002 * plot_height , "white", "white" );
-	
-	// bool Write_SVG_circle( std::ostream& os, double center_x, double center_y, double radius, double stroke_size, 
-    //                    std::string stroke_color , std::string fill_color )
-
-	// write the simulation time to the top of the plot
- 
-	char* szString; 
-	szString = new char [1024]; 
- 
-	int total_cell_count = all_cells->size(); 
- 
-	double temp_time = time; 
-
-	std::string time_label = formatted_minutes_to_DDHHMM( temp_time ); 
- 
-	sprintf( szString , "Current time: %s, z = %3.2f %s", time_label.c_str(), 
-		z_slice , PhysiCell_SVG_options.simulation_space_units.c_str() ); 
-	Write_SVG_text( os, szString, font_size*0.5,  font_size*(.2+1), 
-		font_size, PhysiCell_SVG_options.font_color.c_str() , PhysiCell_SVG_options.font.c_str() );
-	sprintf( szString , "%u agents" , total_cell_count ); 
-	Write_SVG_text( os, szString, font_size*0.5,  font_size*(.2+1+.2+.9), 
-		0.95*font_size, PhysiCell_SVG_options.font_color.c_str() , PhysiCell_SVG_options.font.c_str() );
-	
-	delete [] szString; 
-
-	
-	// add an outer "g" for coordinate transforms 
-	
-	os << " <g id=\"tissue\" " << std::endl 
-	   << "    transform=\"translate(0," << plot_height+top_margin << ") scale(1,-1)\">" << std::endl; 
-	   
-	// prepare to do mesh-based plot (later)
-	
-	double dx_stroma = M.mesh.dx; 
-	double dy_stroma = M.mesh.dy; 
-	
-	os << "  <g id=\"ECM\">" << std::endl; 
-  
-	int ratio = 1; 
-	double voxel_size = dx_stroma / (double) ratio ; 
-  
-	double half_voxel_size = voxel_size / 2.0; 
-	double normalizer = 78.539816339744831 / (voxel_size*voxel_size*voxel_size); 
- 
- // color in the background ECM
-/* 
- if( ECM.TellRows() > 0 )
- {
-  // find the k corresponding to z_slice
-  
-  
-  
-  Vector position; 
-  *position(2) = z_slice; 
-  
-
-  // 25*pi* 5 microns^2 * length (in source) / voxelsize^3
-  
-  for( int j=0; j < ratio*ECM.TellCols() ; j++ )
-  {
-   // *position(1) = *Y_environment(j); 
-   *position(1) = *Y_environment(0) - dy_stroma/2.0 + j*voxel_size + half_voxel_size; 
-   
-   for( int i=0; i < ratio*ECM.TellRows() ; i++ )
-   {
-    // *position(0) = *X_environment(i); 
-    *position(0) = *X_environment(0) - dx_stroma/2.0 + i*voxel_size + half_voxel_size; 
-	
-    double E = evaluate_Matrix3( ECM, X_environment , Y_environment, Z_environment , position );	
-	double BV = normalizer * evaluate_Matrix3( OxygenSourceHD, X_environment , Y_environment, Z_environment , position );
-	if( isnan( BV ) )
-	{ BV = 0.0; }
-
-	vector<string> Colors;
-	Colors = hematoxylin_and_eosin_stroma_coloring( E , BV );
-	Write_SVG_rect( os , *position(0)-half_voxel_size-X_lower , *position(1)-half_voxel_size+top_margin-Y_lower, 
-	voxel_size , voxel_size , 1 , Colors[0], Colors[0] );
-   
-   }
-  }
- 
- }
-*/
-	os << "  </g>" << std::endl; 
- 
-	// Now draw vessels
-
-	/*
-	 std::vector<std::string> VesselColors = hematoxylin_and_eosin_stroma_coloring( 0,1 );
-
-	 os << " <g id=\"BloodVessels\">" << endl; 
-	 extern vector<BloodVesselSegment*> BloodVesselSegments; 
-	 Vector Offset; 
-	 *Offset(0) = X_lower; 
-	 *Offset(1) = Y_lower-top_margin;
-	*/
- 
-
- 
-	// plot intersecting cells 
-	os << "  <g id=\"cells\">" << std::endl; 
-	for( int i=0 ; i < total_cell_count ; i++ )
-	{
-		Cell* pC = (*all_cells)[i]; // global_cell_list[i]; 
-  
-		static std::vector<std::string> Colors; 
-		if( fabs( (pC->position)[2] - z_slice ) < pC->phenotype.geometry.radius )
-		{
-			double r = pC->phenotype.geometry.radius ; 
-			double rn = pC->phenotype.geometry.nuclear_radius ; 
-			double z = fabs( (pC->position)[2] - z_slice) ; 
-   
-			Colors = cell_coloring_function( pC ); 
-
-			os << "   <g id=\"cell" << pC->ID << "\">" << std::endl; 
-  
-			// figure out how much of the cell intersects with z = 0 
-   
-			double plot_radius = sqrt( r*r - z*z ); 
-
-			Write_SVG_circle( os, (pC->position)[0]-X_lower, (pC->position)[1]-Y_lower, 
-				plot_radius , 0.5, Colors[1], Colors[0] ); 
-
-			// plot the nucleus if it, too intersects z = 0;
-			if( fabs(z) < rn && PhysiCell_SVG_options.plot_nuclei == true )
-			{   
-				plot_radius = sqrt( rn*rn - z*z ); 
-			 	Write_SVG_circle( os, (pC->position)[0]-X_lower, (pC->position)[1]-Y_lower, 
-					plot_radius, 0.5, Colors[3],Colors[2]); 
-			}					  
-			os << "   </g>" << std::endl;
-		}
-	}
-	os << "  </g>" << std::endl; 
-
-	// Plot guidelines the circle guides
-	
-	if (line_pattern == "concentric circles")
-	{
-		for (int i=0; i<plot_width/(2*parameters.doubles("ECM_dx")); i++)
-		{
-			double radius = plot_width/2 - (i *parameters.doubles("ECM_dx"));
-			// std::cout<<"Index "<<i<<std::endl;
-			// std::cout<<"Radius "<<radius<<std::endl;
-			Write_SVG_circle( os, plot_width/2, plot_height/2, radius, 0.5, "black" , "none" );
-
-		}
-	}
-
-	else if (line_pattern == "vertical lines")
-	{
-		for (int i=0; i<plot_width/parameters.doubles("ECM_dx"); i++)
-		{
-			double x_line_position = parameters.doubles("ECM_dx")*i;
-			// std::cout<<"Index "<<i<<std::endl;
-			// std::cout<<"X position "<<x_line_position<<std::endl;			
-			Write_SVG_line(os, x_line_position, 0, x_line_position, plot_height, 0.5, "black");
-			// bool Write_SVG_line( std::ostream& os , double start_x, double start_y, double end_x , double end_y, double thickness, 
-            //         std::string stroke_color )
-		}
-	}
-
-	else if (line_pattern == "horizontal lines")
-	{
-		for (int i=0; i<plot_height/parameters.doubles("ECM_dy"); i++)
-		{
-			double y_line_position = parameters.doubles("ECM_dy")*i;
-			// std::cout<<"Index "<<i<<std::endl;
-			// std::cout<<"Y position "<<y_line_position<<std::endl;			
-			Write_SVG_line(os, 0, y_line_position, plot_width, y_line_position, 0.5, "black");
-			// bool Write_SVG_line( std::ostream& os , double start_x, double start_y, double end_x , double end_y, double thickness, 
-            //         std::string stroke_color )
-		}
-	}
-
-	else if (line_pattern == "none") {} // Don't make lines!!!
-
-	else if (line_pattern != "none" || "horizontal lines" || "vertical lines" || "concentric circles")
-	{
-		std::cout<<"Use of this custom SVG output function requires specifying \"none\" or a line pattern" <<std::endl;
-		std::cout<<"Halting: see inputs to custom SVG function \"SVG_plot_custom\"" << std::endl;
-		abort();
-		return;
-
-	}
-
-
-	// plot intersecting BM points
-	/* 
-	 for( int i=0 ; i < BasementMembraneNodes.size() ; i++ )
-	 {
-		// vector<string> Colors = false_cell_coloring( pC ); 
-		BasementMembraneNode* pBMN = BasementMembraneNodes[i]; 
-		double thickness =0.1; 
-		
-		if( fabs( *(pBMN->Position)(2) - z_slice ) < thickness/2.0 ) 
-		{
-		 string bm_color ( "rgb(0,0,0)" );
-		 double r = thickness/2.0; 
-		 double z = fabs( *(pBMN->Position)(2) - z_slice) ; 
-
-		 os << " <g id=\"BMN" << pBMN->ID << "\">" << std::endl; 
-		 Write_SVG_circle( os,*(pBMN->Position)(0)-X_lower, *(pBMN->Position)(1)+top_margin-Y_lower, 10*thickness/2.0 , 0.5 , bm_color , bm_color ); 
-		 os << " </g>" << std::endl;
-		}
-		// pC = pC->pNextCell;
-	 }
-	*/ 
-	
-	// end of the <g ID="tissue">
-	os << " </g>" << std::endl; 
- 
-	// draw a scale bar
- 
-	double bar_margin = 0.025 * plot_height; 
-	double bar_height = 0.01 * plot_height; 
-	double bar_width = PhysiCell_SVG_options.length_bar; 
-	double bar_stroke_width = 0.001 * plot_height; 
-	
-	std::string bar_units = PhysiCell_SVG_options.simulation_space_units; 
-	// convert from micron to mm
-	double temp = bar_width;  
-
-	if( temp > 999 && std::strstr( bar_units.c_str() , PhysiCell_SVG_options.mu.c_str() )   )
-	{
-		temp /= 1000;
-		bar_units = "mm";
-	}
-	// convert from mm to cm 
-	if( temp > 9 && std::strcmp( bar_units.c_str() , "mm" ) == 0 )
-	{
-		temp /= 10; 
-		bar_units = "cm";
-	}
-	
-	szString = new char [1024];
-	sprintf( szString , "%u %s" , (int) round( temp ) , bar_units.c_str() );
- 
-	Write_SVG_rect( os , plot_width - bar_margin - bar_width  , plot_height + top_margin - bar_margin - bar_height , 
-		bar_width , bar_height , 0.002 * plot_height , "rgb(255,255,255)", "rgb(0,0,0)" );
-	Write_SVG_text( os, szString , plot_width - bar_margin - bar_width + 0.25*font_size , 
-		plot_height + top_margin - bar_margin - bar_height - 0.25*font_size , 
-		font_size , PhysiCell_SVG_options.font_color.c_str() , PhysiCell_SVG_options.font.c_str() ); 
-	
-	delete [] szString; 
-
-	// // plot runtime 
-	// szString = new char [1024]; 
-	// RUNTIME_TOC(); 
-	// std::string formatted_stopwatch_value = format_stopwatch_value( runtime_stopwatch_value() );
-	// Write_SVG_text( os, formatted_stopwatch_value.c_str() , bar_margin , top_margin + plot_height - bar_margin , 0.75 * font_size , 
-	// 	PhysiCell_SVG_options.font_color.c_str() , PhysiCell_SVG_options.font.c_str() );
-	// delete [] szString; 
-
-	// draw a box around the plot window
-	Write_SVG_rect( os , 0 , top_margin, plot_width, plot_height , 0.002 * plot_height , "rgb(0,0,0)", "none" );
-	
-	// close the svg tag, close the file
-	Write_SVG_end( os ); 
-	os.close();
- 
-	return; 
-}
-
-void write_ECM_Data_matlab( std::string filename )
-{
-    int number_of_data_entries = ecm.ecm_mesh.voxels.size();
-    int size_of_each_datum = 8;
-	
-	// static int ECM_anisotropy_index = microenvironment.find_density_index( "ECM anisotropy" ); 
-	// static int ECM_density_index = microenvironment.find_density_index( "ECM" ); 
-
-    FILE* fp = write_matlab_header( size_of_each_datum, number_of_data_entries,  filename, "ECM_Data" );  // Note - the size of datum needs to correspond exaectly to the lines of output or there is an error upon importing.
-
-    for( int i=0; i < number_of_data_entries ; i++ )
-    {
-	    fwrite( (char*) &( ecm.ecm_mesh.voxels[i].center[0] ) , sizeof(double) , 1 , fp ); // 1
-        fwrite( (char*) &( ecm.ecm_mesh.voxels[i].center[1] ) , sizeof(double) , 1 , fp ); // 2
-        fwrite( (char*) &( ecm.ecm_mesh.voxels[i].center[2] ) , sizeof(double) , 1 , fp ); //3
-		fwrite( (char*) &( ecm.ecm_voxels[i].anisotropy), sizeof(double) , 1 , fp ); // 4
-        fwrite( (char*) &( ecm.ecm_voxels[i].density), sizeof(double) , 1 , fp ); // 5
-        fwrite( (char*) &( ecm.ecm_voxels[i].ecm_fiber_alignment[0]), sizeof(double) , 1 , fp ); // 6
-        fwrite( (char*) &( ecm.ecm_voxels[i].ecm_fiber_alignment[1]), sizeof(double) , 1 , fp ); // 7
-        fwrite( (char*) &( ecm.ecm_voxels[i].ecm_fiber_alignment[2]), sizeof(double) , 1 , fp ); // 8
-
-		// This will only work if the diffusion and ECM meshes are the same size. Commenting out for actualrunning. To do a direct comparison, leave them in and change length. Will have to change the vizualization to get this from the regular BioFVM outputs.
-
-		// fwrite( (char*) &( microenvironment.gradient_vector(i)[0][0]), sizeof(double) , 1 , fp ); // 9
-		// fwrite( (char*) &( microenvironment.gradient_vector(i)[0][1]), sizeof(double) , 1 , fp ); // 10
-		// fwrite( (char*) &( microenvironment.gradient_vector(i)[0][2]), sizeof(double) , 1 , fp ); // 11
-
-    }
-    fclose( fp );
     return;
 }
