+++ custom_modules/fibrosis.h
@@ -133,6 +133,6 @@ void change_migration_bias_vector_ecm(Cell* pCell , Phenotype& phenotype , doubl
 void run_biotransport( double t_max );
 void alter_cell_uptake_secretion_saturation ( void );
 void set_cell_motility_vectors(void); // Runs update_migration_bias for each cell present in a simulation
-void SVG_plot_custom( std::string filename , Microenvironment& M, double z_slice , double time, std::vector<std::string> (*cell_coloring_function)(Cell*), std::string line_pattern );
-void write_ECM_Data_matlab( std::string filename );
+// void SVG_plot_custom( std::string filename , Microenvironment& M, double z_slice , double time, std::vector<std::string> (*cell_coloring_function)(Cell*), std::string line_pattern );
+// void write_ECM_Data_matlab( std::string filename );
 double sign_function (double number);
